Clean up announce
Summary:
Move announce endpoint query args into a formal body. Avoids a bunch of unnecessary
parsing / converting of types and just relies on JSON decoding.

The benefit here is we no longer pass around stringified InfoHashes or PeerIDs, instead
using the formal type.

Reviewers: #kraken, yiran

Reviewed By: #kraken, yiran

Subscribers: jenkins

Differential Revision: https://code.uberinternal.com/D1573325

+++ tracker/peerhandoutpolicy/ipv4netmask_peer_priority_policy.go
@@ -43,7 +43,7 @@ func (p *IPv4NetmaskPeerPriorityPolicy) AssignPeerPriority(
 		peer.Priority = calcPriority(predicates)
 		if peer.Origin {
 			// Origin peers always have worse priority than regular peers.
-			peer.Priority += int64(len(predicates))
+			peer.Priority += len(predicates)
 		}
 	}
